==> code/01_plot_aesthetics.R <==
#
# manual scales
#

  alpha_global <- 0.5
  
  colors <- c(
    c1 = "#53CCED",
    m37 = "#E018F2",
    w2 = "#ED5653",
    a45 = "#E3AD4F",
    st = "#47E645",
    "FALSE" = "#1155cc", # blue
    "TRUE" = "#cc1111", # red
    "absent" = "#1155cc", # blue
    "present" = "#cc1111" # red
  )
  
  linetypes <- c(
    "FALSE" = "dotted",
    "TRUE" = "solid",
    
    "absent" = "dotted",
    "present" = "solid",

    "7" = "dashed",
    "14" = "solid",
    
    "one" = "dashed",
    "two" = "solid"
  )
  
  shapes <- c(
    "FALSE" = 1,
    "TRUE" = 16
  )
  
  

#
# Sizes
#

  # geoms
    size_point <- 8
    size_path <- 1
    
    
  # plot elements
    thickness_axisLine <- 2
    
  # text
    size_axisTitle <- 18
    size_axisText <- 12

#
# Other
#
  
  fontChoice <- "Proxima Nova"
